CLN: specify lookup dict for experiment
