+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 # Since the point of the package is reproducibility, incl. all dev
 # dependencies
 dependencies = [
-  "mitosis @ git+https://github.com/Jacob-Stevens-Haas/mitosis@main",
+  "mitosis @ git+https://github.com/Jacob-Stevens-Haas/mitosis@0.3.0",
   "derivative @ git+https://github.com/Jacob-Stevens-Haas/derivative@hyperparams",
   "pysindy[cvxpy,miosr] @ git+https://github.com/dynamicslab/pysindy@master",
   "kalman @ git+https://github.com/Jacob-Stevens-Haas/kalman@0.1.0",
